// packages/stable/core/src/types/enum/theme-color-type.ts
/**
 * Copyright 2023-present DreamNum Co., Ltd.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/**
 * Theme color type enum
 */
export enum ThemeColorType {
    /**
     * TEXT
     */
    DARK1,
    /**
     * BACKGROUND
     */
    LIGHT1,
    DARK2,
    LIGHT2,
    ACCENT1,
    ACCENT2,
    ACCENT3,
    ACCENT4,
    ACCENT5,
    ACCENT6,
    /**
     * LINK
     */
    HYPERLINK,
    FOLLOWED_HYPERLINK,
}

// TODO: get color list to color picker
// TODO: 取色

export enum ThemeColors {
    OFFICE = 'Office',
    OFFICE_2007_2010 = 'Office 2007-2010',
    GRAYSCALE = 'Grayscale',
    BLUE_WARM = 'Blue Warm',
    BLUE = 'Blue',
    BLUE_II = 'Blue II',
    BLUE_GREEN = 'Blue Green',
    GREEN = 'Green',
    GREEN_YELLOW = 'Green Yellow',
    YELLOW = 'Yellow',
    YELLOW_ORANGE = 'Yellow Orange',
    ORANGE = 'Orange',
    ORANGE_RED = 'Orange Red',
    RED_ORANGE = 'Red Orange',
    RED = 'Red',
    RED_VIOLET = 'Red Violet',
    VIOLET = 'Violet',
    VIOLET_II = 'Violet II',
    MEDIAN = 'Median',
    PAPER = 'Paper',
    MARQUEE = 'Marquee',
    SLIPSTREAM = 'Slipstream',
    Aspect = 'Aspect',
}